feat(compare): compare two csv files with compare section
Starting small with a simple compare-2-csv's use case, this has already been a great testing ground for getting the right data from the backend.

One of the biggest challenges is the schema. Inferring the schena from diff output is a fools errand, especially when schemas are being inferred on the backend for diff creation in the first place